refactor: move components to a wrapper folder
