+++ components/Card/index.tsx
@@ -1,0 +1,1 @@
+export * from './Card'
