+++ components/GradientText/GradientText.spec.tsx
@@ -1,4 +1,4 @@
-import { render, screen } from '../test'
+import { render, screen } from '../../test'
 import { describe, it, expect } from 'vitest'
 
 import { GradientText } from './GradientText'
