+++ components/GradientText/index.tsx
@@ -1,0 +1,1 @@
+export * from './GradientText'
